% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_calculations_mz.R
\name{calculate_rabd}
\alias{calculate_rabd}
\title{Calculate Relative Absorption Band Depth (RABD)}
\usage{
calculate_rabd(raster, edges, trough, rabd_name)
}
\arguments{
\item{raster}{a terra SpatRaster of normalized capture data}

\item{edges}{a numeric vector of two for the wide calculation window}

\item{trough}{a character vector of wavelenght to look for trough}

\item{rabd_name}{a character, name of calculated RABD}
}
\value{
a terra SpatRaster with one layer with calculated RABD values
}
\description{
Calculate Relative Absorption Band Depth (RABD)
}
\section{RABD655−680max = ( X × R590 + Y × R730 X+Y ) /R655−680min}{
Denominator: Find the lowest reflectance between 655 and 680
Numenator: find reflectance at 590, find reflectance at 730
Find how many bands there are between trough minimum and 730
Find how many bands there are between trough minimum and 590
}

