% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_calculations_mz.R
\name{calculate_rmean}
\alias{calculate_rmean}
\title{calculate rMean}
\usage{
calculate_rmean(raster)
}
\arguments{
\item{raster}{a terra SpatRaster of normalized capture data}
}
\value{
a terra SpatRaster with one layer with calculated rMean values
}
\description{
calculate mean reflectance from all layers for given pixel
}
