% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralCalculationsNick.R
\name{calculateBandRatio}
\alias{calculateBandRatio}
\title{Calculate a Band Ratio}
\usage{
calculateBandRatio(normData, normWavelengths, tol = 1, top = 570, bot = 630)
}
\arguments{
\item{normData}{Normalized spectral data (the output of \code{specimR::calculateMeanRows()})}

\item{normWavelengths}{A vector of wavelengths (the output of \code{specimR::calculateMeanRows()})}

\item{tol}{Tolerance for distance from the requested wavelength. If no values are present within the tolerance, this returns an error. If multiple are present, this returns the average of those values.}

\item{top}{the wavelength in the top of the ratio (the numerator) (default = 570)}

\item{bot}{the wavelength in the top of the ratio (the denominator) (default = 630)}
}
\value{
a depth series of the band ratio
}
\description{
Calculate a Band Ratio
}
