% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralCalculationsNick.R
\name{calculateIndices}
\alias{calculateIndices}
\title{calculate a suite of spectral indices on normalized data}
\usage{
calculateIndices(
  normalized,
  indices = c("RABD660", "RABD660670", "RABD845", "R570R630", "R590R690"),
  tol = 1,
  smooth.win = round(0.2/normalized$cmPerPixel)
)
}
\arguments{
\item{normalized}{Normalized spectral data (the output of `specimR::getNormValues()`)}

\item{indices}{a character vector of the requested indices. Current options are: "RABD660","RABD845","R570R630"and "R590R690" (default = c("RABD660","RABD845","R570R630","R590R690"))}

\item{tol}{the tolerance to use when matching wavelengths (in nm)}
}
\value{
a data.frame with depth and the requested indices
}
\description{
calculate a suite of spectral indices on normalized data
}
