% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_drive_depth_mz.R
\name{composite_drive}
\alias{composite_drive}
\title{Create composite drive from multiple rois}
\usage{
composite_drive(depth, indices)
}
\arguments{
\item{depth}{a named list of tibbles with depths by roi.}

\item{indices}{a named list of tibble with spectral indices}
}
\value{
a tibble with depths translated for each drive.
}
\description{
Create composite drive from multiple rois
}
