% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositePlotting.R
\name{compositeSections}
\alias{compositeSections}
\title{Title}
\usage{
compositeSections(dirs = NA, out.dir = NA, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=plotCompositeSpectralDashboard]{plotCompositeSpectralDashboard}}
  \describe{
    \item{\code{normList}}{a list of the normalized output lists from normalize()}
    \item{\code{coreTable}}{core depth table}
    \item{\code{indices}}{a list of spectralIndices tibbles}
    \item{\code{output.dpi}}{}
  }}
}
\value{

}
\description{
Title
}
