% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_mz.R
\name{create_normalized_raster}
\alias{create_normalized_raster}
\title{Raster normalization}
\usage{
create_normalized_raster(
  capture = capture,
  whiteref = whiteref,
  darkref = darkref,
  fun = normalization,
  ...
)
}
\arguments{
\item{capture}{a terra SpatRaster of captured data}

\item{whiteref}{a terra SpatRaster of the white reference matching capture extent}

\item{darkref}{a terra SpatRaster of the dark reference matching capture extent}

\item{fun}{function to apply: normalization}
}
\value{
a normalized terra SpatRaster of the capture
}
\description{
apply normalization function over the combination of capture and reference SpatRasters using terra spatial dataset
}
