% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_mz.R
\name{create_reference_raster}
\alias{create_reference_raster}
\title{Create reference raster}
\usage{
create_reference_raster(raster, roi, ref_type)
}
\arguments{
\item{raster}{a terra SpatRaster of the captured reference}

\item{roi}{Region Of Interest: extent to match data raster}
}
\value{
a terra SpatRaster of reference matching the data raster extent
}
\description{
Creating reference SpatRaster covering core extent
Create one mean reference row SpatRaster by averaging data every column by aggregation
Create reference SpatRaster matching capture SpatRaster extent by disaggregation
}
