% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeDepth.R
\name{createCoreDepthTable}
\alias{createCoreDepthTable}
\title{Create a core-depth table for depth compositing}
\usage{
createCoreDepthTable(cores)
}
\arguments{
\item{cores}{A named list of coreTables export from normalize()}
}
\value{

}
\description{
Create a core-depth table for depth compositing
}
