% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageProcessing.R
\name{createImages}
\alias{createImages}
\title{Create image from reflectance standardized data}
\usage{
createImages(
  bigRoi = NA,
  directory = NA,
  wavelengths = c(630, 532, 465),
  image.output.dir = NA,
  stretch.method = "full",
  stretch.fun = "linear2pct",
  guessBigRoi = FALSE
)
}
\arguments{
\item{stretch.fun}{}
}
\description{
Create image from reflectance standardized data
}
