% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{defineRGB}
\alias{defineRGB}
\title{choose the layers to use for plotting red, green, and blue}
\usage{
defineRGB(wavelengths)
}
\arguments{
\item{wavelengths}{}
}
\value{
best matched layers for the three colors
}
\description{
choose the layers to use for plotting red, green, and blue
}
