% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_calculations_mz.R
\name{extract_profile}
\alias{extract_profile}
\title{Extract average proxy profile}
\usage{
extract_profile(raster)
}
\arguments{
\item{raster}{a terra SpatRaster with one layer with calculated values}
}
\value{
a data frame with XY coordinates and averaged proxy values
}
\description{
Extract average proxy profile
}
