% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_mz.R
\name{filter_savgol}
\alias{filter_savgol}
\title{Apply a Savitzky-Golay smoothing filter}
\usage{
filter_savgol(raster, p = 3, n = p + 3 - p\%\%2, m = 0, ts = 1)
}
\arguments{
\item{raster}{a terra SpatRaster of normalized data}

\item{p}{filter order.}

\item{n}{filter length (must be odd).}

\item{m}{return the m-th derivative of the filter coefficients.}

\item{ts}{time scaling factor.}
}
\value{
A filtered terra SpatRaster.
}
\description{
Smooth data with a Savitzky-Golay smoothing filter using \code{\link[signal]{sgolayfilt}}.
}
