% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_spectra.R
\name{fullSpectra}
\alias{fullSpectra}
\title{Normalize a hyperspectral image}
\usage{
fullSpectra(
  directory = NA,
  length.out = NA,
  wavelengths = NA,
  cmPerPixel = NA,
  roi = NA,
  output.dir = NA,
  corename = NA,
  chunk.top = 0,
  chunk.bot = 3,
  chunk.step = 0.25
)
}
\arguments{
\item{directory}{path to Specim core folder}

\item{cmPerPixel}{optionally specify the number of cm per pixel. Will determine interactively if NA. (default = NA)}

\item{roi}{rasterExtent object defining region of interest. NA will allow you to choose interactively (default = NA)}

\item{output.dir}{optionally save output raster by specifyign the path}

\item{corename}{optionally choose the length of the core}

\item{spectra}{which spectra to normalize}
}
\value{
a normalized hyperspectral image
}
\description{
Normalize a hyperspectral image
}
