% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_spectra.R
\name{fullSpectraWholeROI}
\alias{fullSpectraWholeROI}
\title{Normalize a hyperspectral image}
\usage{
fullSpectraWholeROI(directory = NA, roi = NA, output.dir = NA, corename = NA)
}
\arguments{
\item{directory}{path to Specim core folder}

\item{roi}{rasterExtent object defining region of interest. NA will allow you to choose interactively (default = NA)}

\item{output.dir}{optionally save output raster by specifyign the path}

\item{corename}{optionally choose the length of the core}

\item{cmPerPixel}{optionally specify the number of cm per pixel. Will determine interactively if NA. (default = NA)}

\item{spectra}{which spectra to normalize}
}
\value{
a normalized hyperspectral image
}
\description{
Normalize a hyperspectral image
}
