% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-operations_mz.R
\name{get_paths}
\alias{get_paths}
\title{Get paths to files}
\usage{
get_paths(directory = NA)
}
\arguments{
\item{directory}{directory to search for files}
}
\value{
a list of paths pointing to the overview and capture files
}
\description{
search for an overview file (\emph{.png) and three files of captured data
core capture, white and dark references (}.raw).
}
