% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralCalculationsNick.R
\name{getNormValues}
\alias{getNormValues}
\title{Transpose normalized dataset and get normalized values}
\usage{
getNormValues(normalized)
}
\arguments{
\item{normalized}{the output of specimR::normalize()}
}
\value{
matrix of normalized values averaged across the core
}
\description{
Transpose normalized dataset and get normalized values
}
