% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralCalculationsNick.R
\name{getNormWavelengthData}
\alias{getNormWavelengthData}
\title{Extract the depth series for a specific wavelength}
\usage{
getNormWavelengthData(
  normData,
  normWavelengths,
  wavelengthToGet,
  tol = 1,
  agg.fun = "mean"
)
}
\arguments{
\item{normData}{Normalized spectral data (the output of `specimR::getNormValues()`)}

\item{normWavelengths}{A vector of wavelengths (the output of `specimR::getNormValues()`)}

\item{wavelengthToGet}{What wavelength to do you want to extract}

\item{tol}{Tolerance for distance from the requested wavelength. If no values are present within the tolerance, this returns an error. If multiple are present, this returns the average of those values.}
}
\value{
The depth sequence of normalized reflectance values for the requested wavelength.
}
\description{
Extract the depth series for a specific wavelength
}
