% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralCalculationsNick.R
\name{getNormWavelengths}
\alias{getNormWavelengths}
\title{get the wavelengths present in the normalized matrix}
\usage{
getNormWavelengths(normValues)
}
\arguments{
\item{normValues}{the output of \code{specimR::calculateMeanRows()}}
}
\value{
a vector of wavelengths
}
\description{
get the wavelengths present in the normalized matrix
}
