% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralCalculationsNick.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Title}
\usage{
makeHeatmap(
  normalized,
  index = "RABD660",
  tol = 1,
  smooth = TRUE,
  smooth.sigma = 3,
  smooth.n = 7
)
}
\arguments{
\item{smooth.n}{}
}
\value{

}
\description{
Title
}
