% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a hyperspectral image}
\usage{
normalize(
  directory = NA,
  length = NA,
  spectra = c(615, 660, 670),
  roi = NA,
  tif.path.to.write = NA
)
}
\arguments{
\item{wavelengths}{a vector of wavelengths to extract from the hyperspectral image (cho)}

\item{data.file}{optionally specify the path to the hyperspectral .raw image}
}
\value{
a normalized hyperspectral image
}
\description{
Normalize a hyperspectral image
}
