% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_mz.R
\name{plot_profile_proxy}
\alias{plot_profile_proxy}
\title{Line plots of calculated proxies}
\usage{
plot_profile_proxy(data, hsi_index, .palette)
}
\arguments{
\item{hsi_index}{a character indicating hyperspectral index layer to plot.}

\item{.palette}{a character indicating one of color palettes of choice. One of: "red", "orange", "yellow", "green", "cyan", "blue", "purple", "magenta".}

\item{raster}{a SpatRaster with calculated hyperspectral indices and RGB layers.}
}
\value{
a plot with color map of selected hyperspectral index.
}
\description{
Line plots of calculated proxies
}
