% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_mz.R
\name{plot_raster_proxy}
\alias{plot_raster_proxy}
\title{Spatial map plots of calculated proxies}
\usage{
plot_raster_proxy(raster, hsi_index, .palette)
}
\arguments{
\item{raster}{a SpatRaster with calculated hyperspectral indices and RGB layers.}

\item{hsi_index}{a character indicating hyperspectral index layer to plot.}

\item{.palette}{a character indicating one of color palettes of choice. One of: "red", "orange", "yellow", "green", "cyan", "blue", "purple", "magenta".}
}
\value{
a plot with color map of selected hyperspectral index.
}
\description{
Spatial map plots of calculated proxies
}
