% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositePlotting.R
\name{plotCompositeSpectralDashboard}
\alias{plotCompositeSpectralDashboard}
\title{Plot composite spectral dashboard}
\usage{
plotCompositeSpectralDashboard(
  normList,
  coreTable,
  processed.image.dir = file.path(normList[[1]]$outputDir, "photos"),
  index.name = "RABD660",
  depth.label = "Depth (cm)",
  core.width = 4,
  plot.width = 8,
  page.width = 10,
  y.tick.interval = 5,
  page.units = "cm",
  tol = 1,
  output.file.path = NA,
  output.dpi = 600
)
}
\arguments{
\item{tol}{}
}
\value{

}
\description{
Plot composite spectral dashboard
}
