% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{plotSpectralOutput}
\alias{plotSpectralOutput}
\title{Plot spectral output}
\usage{
plotSpectralOutput(
  normalized,
  indices = c("RABD615", "RABD660670", "RABD845", "R570R630", "R590R690"),
  file.type = "png",
  fig.width = 20,
  fig.length = NA,
  fig.units = "in",
  ...
)
}
\arguments{
\item{normalized}{The output of normalize()}

\item{indices}{Which indices to plot? see ?calculateIndices for options}

\item{file.type}{Export file type (default = "png")}

\item{fig.width}{Width of the overview plot in fig.units (individual plots will be half as wide) (default = 20)}

\item{fig.length}{Length of the overview plot in the fig.units, if NA, will use core length (default = NA)}

\item{fig.units}{Units for fig dimensions (default = "in"s)}

\item{...}{
  Arguments passed on to \code{\link[=plotSpectralDashboard]{plotSpectralDashboard}}
  \describe{
    \item{\code{output.dpi}}{}
  }}
}
\description{
Plot spectral output
}
