% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{processReferenceLarge}
\alias{processReferenceLarge}
\title{Perform reference processing prior to normalizations for large numbers of spectra}
\usage{
processReferenceLarge(reference, stripe, spectra)
}
\arguments{
\item{reference}{a rasterBrick object to process}

\item{stripe}{a raster whose extent you want to mirror}

\item{spectra}{an index of spectra to subset}
}
\value{
a standardized rasterBrick the same size and thickness as stripe
}
\description{
Perform reference processing prior to normalizations for large numbers of spectra
}
