% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_mz.R
\name{raster_crop}
\alias{raster_crop}
\title{Raster crop}
\usage{
raster_crop(raster, type, roi, ...)
}
\arguments{
\item{raster}{a terra SpatRaster to be cropped}

\item{type}{either data raster or reference raster}

\item{roi}{Region Of Interest: cropping extent}
}
\value{
a terra SpatRaster cropped to ROI
}
\description{
Crop SpatRaster to large ROI (entire core)
For capture (core) SpatRaster use full extent
For reference (white and dark) SpatRaster use only x-direction
}
