% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralCalculationsNick.R
\name{rasterBandRatio}
\alias{rasterBandRatio}
\title{calculate band ratio over a raster}
\usage{
rasterBandRatio(normalized, top = 570, bot = 630, tol = 1)
}
\arguments{
\item{top}{the wavelength in the top of the ratio (the numerator) (default = 570)}

\item{bot}{the wavelength in the top of the ratio (the denominator) (default = 630)}

\item{tol}{tolerance for finding the nearest wavelength (default = 1)}

\item{trough}{wavelength for the RABD trough}

\item{edges}{a 2-element vector marking the edges of the trough}
}
\value{
a raster object with the spatial band ratio
}
\description{
calculate band ratio over a raster
}
