% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralCalculationsNick.R
\name{rasterRABD}
\alias{rasterRABD}
\title{calculate RABD over a raster}
\usage{
rasterRABD(
  normalized,
  trough = 660,
  edges = c(590, 730),
  tol = 1,
  trough.agg.fun = "min",
  edge.agg.fun = "mean"
)
}
\arguments{
\item{normalized}{output list of specimR::normalize()}

\item{trough}{wavelength for the RABD trough}

\item{edges}{a 2-element vector marking the edges of the trough}

\item{tol}{tolerance for finding the nearest wavelength (default = 1)}
}
\value{

}
\description{
calculate RABD over a raster
}
