% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_drive_depth_mz.R
\name{read_depth}
\alias{read_depth}
\title{Read depth tables data from directory}
\usage{
read_depth(path)
}
\arguments{
\item{path}{path to the root core directory containing drives.}
}
\value{
named list of depth tables from drives within the core.
}
\description{
Read depth tables data from directory
}
