% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_drive_depth_mz.R
\name{read_hsi}
\alias{read_hsi}
\title{Read spectral indices data from directory}
\usage{
read_hsi(path)
}
\arguments{
\item{path}{path to the root core directory containing the drives.}
}
\value{
named list of spectral indices from drives within the core.
}
\description{
Read spectral indices data from directory
}
