% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_calculations_mz.R
\name{remove_continuum}
\alias{remove_continuum}
\title{Remove continuum from spectrum}
\usage{
remove_continuum(raster, ...)
}
\arguments{
\item{raster}{a terra SpatRaster of normalized capture data}
}
\value{
a terra SpatRaster of normalized capture data with continuum removed
}
\description{
Remove continuum from spectrum
}
