% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{spectralWorkflow}
\alias{spectralWorkflow}
\title{Wrapper function that goes through standardization, index calculation and visualization}
\usage{
spectralWorkflow(
  directory = NA,
  indices = c("RABD615", "RABD660670", "RABD845", "R570R630", "R590R690"),
  overall.page.width = 30,
  individual.page.width = 10,
  plot.width = 8,
  core.width = 4,
  image.wavelengths = c(630, 532, 465),
  predict.image.roi = FALSE,
  imageRoi = NA,
  output.dir = NA,
  clickDepths = NA,
  overall.width = NA,
  individual.width = NA,
  width.mult = NA,
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=normalize]{normalize}}
  \describe{
    \item{\code{directory}}{path to Specim core folder}
    \item{\code{cmPerPixel}}{optionally specify the number of cm per pixel. Will determine interactively if NA. (default = NA)}
    \item{\code{roi}}{rasterExtent object defining region of interest. NA will allow you to choose interactively (default = NA)}
    \item{\code{output.dir}}{optionally save output raster by specifyign the path}
    \item{\code{corename}}{optionally choose the length of the core}
  }}
}
\value{

}
\description{
Wrapper function that goes through standardization, index calculation and visualization
}
