% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralCalculationsNick.R
\name{calculateRABD}
\alias{calculateRABD}
\title{Calculate Relative Absorption Band Depth (RABD)}
\usage{
calculateRABD(
  normData,
  normWavelengths,
  tol = 5,
  trough = 665,
  edges = c(590, 730),
  trough.agg.fun = "min"
)
}
\arguments{
\item{normData}{Normalized spectral data (the output of \code{specimR::calculateMeanRows()})}

\item{normWavelengths}{A vector of wavelengths (the output of \code{specimR::calculateMeanRows()})}

\item{tol}{Tolerance for distance from the requested wavelength. If no values are present within the tolerance, this returns an error. If multiple are present, this returns the average of those values.}

\item{trough}{the wavelength for the RABD trough (default = 660)}

\item{edges}{the edges (as a 2-element vector) of the RABD trough (default = c(590,730))}
}
\value{
a depth series of the RABD index
}
\description{
Calculate Relative Absorption Band Depth (RABD)
}
