% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_mz.R
\name{median_filtering}
\alias{median_filtering}
\title{Smooth raster with focal median}
\usage{
median_filtering(caoture = capture, window = 3)
}
\arguments{
\item{window}{focal window size, default is 3}

\item{capture}{a terra SpatRaster of captured data}
}
\value{
smoothed SpatRaster
}
\description{
Smooth raster with focal median
}
