% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositePlotting.R
\name{plotCompositeSpectralDashboard}
\alias{plotCompositeSpectralDashboard}
\title{Plot composite spectral dashboard}
\usage{
plotCompositeSpectralDashboard(
  normList,
  coreTable,
  indices = NA,
  processed.image.dirs = NA,
  index.name = "RABD660670",
  depth.label = "Depth (cm)",
  core.width = 4,
  plot.width = 8,
  page.width = 10,
  y.tick.interval = 5,
  page.units = "cm",
  tol = 1,
  output.file.path = NA,
  output.dpi = 600
)
}
\arguments{
\item{normList}{a list of the normalized output lists from normalize()}

\item{coreTable}{core depth table}

\item{indices}{a list of spectralIndices tibbles}

\item{output.dpi}{}
}
\value{

}
\description{
Plot composite spectral dashboard
}
