% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickViz.R
\name{plotSpectralDashboard}
\alias{plotSpectralDashboard}
\title{Title}
\usage{
plotSpectralDashboard(
  normalized,
  ind,
  processed.image.dir = file.path(normalized$outputDir, "photos"),
  index.name = "RABD660",
  depth.label = "Depth (cm)",
  core.width = 4,
  plot.width = 8,
  page.width = 10,
  page.width.multiplier = 1.5,
  page.length.multiplier = 3,
  y.tick.interval = 5,
  page.units = "cm",
  tol = 1,
  output.file.path = NA,
  output.dpi = 600
)
}
\arguments{
\item{output.dpi}{}
}
\description{
Title
}
