% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickViz.R
\name{plotSpectralDashboard}
\alias{plotSpectralDashboard}
\title{Title}
\usage{
plotSpectralDashboard(
  normalized,
  ind,
  index.name = "RABD660",
  depth.label = "Depth (cm)",
  width.mult = 1,
  plot.width = 3,
  tol = 1
)
}
\arguments{
\item{tol}{}
}
\value{

}
\description{
Title
}
