% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_mz.R
\name{prepare_core}
\alias{prepare_core}
\title{Prepare core based on shiny output}
\usage{
prepare_core(
  path = choices$directory,
  .normalize = choices$analysisOptions$normalize
)
}
\arguments{
\item{.normalize}{}
}
\description{
Prepare core based on shiny output
}
